+++ functions/src/api/Api.ts
@@ -1,4 +1,5 @@
 export type SlackParams = {
   slackTeamId: string
 }
+
 export type ApiStatus = 'ok' | 'error'
